% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airGRccia-package.R
\docType{package}
\name{airGRccia-package}
\alias{airGRccia}
\alias{airGRccia-package}
\title{airGRccia: Climate Change Impact and Adaptation with airGRiwrm}
\description{
This package provides tools for studying climate Change Impact and Adaptation on a watershed modeled with airGRiwrm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umr-g-eau.pages.mia.inra.fr/airgrccia}
  \item \url{https://forgemia.inra.fr/umr-g-eau/airgrccia}
  \item Report bugs at \url{https://forgemia.inra.fr/umr-g-eau/airgrccia/-/issues}
}

}
\author{
\strong{Maintainer}: David Dorchies \email{david.dorchies@inrae.fr} (\href{https://orcid.org/0000-0002-6595-7984}{ORCID})

}
\keyword{internal}
