% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_extract_cells.R
\name{climate_extract_cells}
\alias{climate_extract_cells}
\title{Extract cell values from a climate dataset}
\usage{
climate_extract_cells(
  cells,
  s = climate_read(GCM, RCM, BC, RCP, path),
  GCM,
  RCM,
  BC,
  RCP,
  path = getwd()
)
}
\arguments{
\item{cells}{\link{integer} Number of SAFRAN cells}

\item{s}{A stars object read by \link{climate_read}}

\item{GCM}{\link{character} Selected GCM}

\item{RCM}{\link{character} Selected RCM}

\item{BC}{\link{character} Selected Bias Correction}

\item{RCP}{The chosen RCP (either "rcp26", "rcp45", or "rcp85")}

\item{path}{\link{character} Local path of the downloaded file}
}
\value{
A \link{list} with one item \code{date} and one item by variable present in the
stars object \code{s}. The item \code{date} is a vector and each other item is a data.frame
with the SAFRAN cell number as column and one row by time step.
}
\description{
Extract cell values from a climate dataset
}
\details{
SAFRAN cell numbers are compared to numbers from the table available by typing:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{readr::read_tsv(system.file("extdata", "mailles_safran_drias.tsv", package = "airGRccia"),
                comment = "#",
               show_col_types = FALSE)
}\if{html}{\out{</div>}}
}
