% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_read.R
\name{climate_is_available}
\alias{climate_is_available}
\title{Check that climate scenario netCDF files exist}
\usage{
climate_is_available(GCM, RCM, BC, RCP, path = getwd())
}
\arguments{
\item{GCM}{\link{character} Selected GCM}

\item{RCM}{\link{character} Selected RCM}

\item{BC}{\link{character} Selected Bias Correction}

\item{RCP}{The chosen RCP (either "rcp26", "rcp45", or "rcp85")}

\item{path}{\link{character} Local path of the downloaded file}
}
\value{
A \link{logical} with an attribute "details" specifying which files are missing
}
\description{
Check that climate scenario netCDF files exist
}
