% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_read.R
\name{climate_read}
\alias{climate_read}
\title{Read climate data from NetCDF files}
\usage{
climate_read(GCM, RCM, BC, RCP, path = getwd())
}
\arguments{
\item{GCM}{\link{character} Selected GCM}

\item{RCM}{\link{character} Selected RCM}

\item{BC}{\link{character} Selected Bias Correction}

\item{RCP}{The chosen RCP (either "rcp26", "rcp45", or "rcp85")}

\item{path}{\link{character} Local path of the downloaded file}
}
\value{
A stars object
}
\description{
Read and concatenate historical and RCP climate netCDF files provided by
\code{drias_crop_agg_save}.
}
