% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasinsObs.R
\name{CreateBasinsObs}
\alias{CreateBasinsObs}
\title{Create a \code{BasinsObs} object (Hydroclimatic time series)}
\usage{
CreateBasinsObs(
  s_sb,
  Qobs = NULL,
  Qinf = NULL,
  Qrelease = NULL,
  measures = c("Precip", "PotEvap", "TempMean")
)
}
\arguments{
\item{s_sb}{Climate data (See \link{get_climate_safran})}

\item{Qobs}{Observed flows (See \link{get_hubeau_flows})}

\item{Qinf}{(optional) \link{matrix} or \link{data.frame} of \link{numeric} containing
observed flows. It must be provided only for nodes of type "Direct
injection" and "Diversion". See \link[airGRiwrm]{CreateGRiwrm} for
details about these node types. Unit is [mm per time step] for nodes
with an area, and [m³ per time step] for nodes with \code{area=NA}.
Column names correspond to node IDs. Negative flows are abstracted from
the model and positive flows are injected to the model}

\item{Qrelease}{(optional) \link{matrix} or \link{data.frame} of \link{numeric} containing
release flows by nodes using the model \code{RunModel_Reservoir} [m3 per
time step]}

\item{measures}{}
}
\value{
A \link{list} containing an item\code{DatesR}, vector of dates corresponding
to the time dimension of the data, following by \link{data.frame}s for each measure
including climate data ("Precip", "PotEvap", "TempMean"), and observed
flows ("Q").
}
\description{
Create a \code{BasinsObs} object (Hydroclimatic time series)
}
