% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drias_crop_aggregate.R
\name{drias_crop_aggregate}
\alias{drias_crop_aggregate}
\alias{drias_crop_agg_save}
\title{Create a stars object for one DRIAS projection}
\usage{
drias_crop_aggregate(
  GCM,
  RCM,
  BC,
  scenario,
  variables = cfg$climato$drias$variables,
  bbox,
  drias = drias_list(cfg = cfg, with_variables = TRUE),
  path = tempdir(),
  cfg = loadConfig()
)

drias_crop_agg_save(
  GCM,
  RCM,
  BC,
  scenario,
  variables,
  bbox,
  path = getwd(),
  cfg = loadConfig()
)
}
\arguments{
\item{GCM}{\link{character} Selected GCM}

\item{RCM}{\link{character} Selected RCM}

\item{BC}{\link{character} Selected Bias Correction}

\item{scenario}{\link{character} Selected scenario ("historical", "rcp2.6", "rcp4.5", or "rcp8.5")}

\item{variables}{\link{character} \link{vector} of variables to merge}

\item{bbox}{\link{vector} of bounding box coordinates in the form
\code{c(xmin, ymin, xmax, ymax)} in the Lambert II CRS ("EPSG:27572").}

\item{drias}{\link{data.frame} containing the DRIAS URLs and metadata.}

\item{path}{\link{character} Local path of the downloaded file}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for details}
}
\value{
A stars object cropped to the study area with all selected variables for the
selected GCM/RCM/scenario.
}
\description{
DRIAS data are downloaded for each requested variable, the study area is
cropped and all variables are merged into one stars object.
}
\details{
URLs for downloading NetCDF files are picked from the \code{drias_GCM_RCM_scenario_variables.tsv}
file stored on Nextcloud and created by \link{drias_list_remote}.
}
\examples{
\dontrun{
nc <- drias_crop_aggregate("CNRM-CERFACS-CNRM-CM5",
                           "CNRM-ALADIN63",
                           "ADAMONT",
                           "historical",
                           c("prtot", "evspsblpot"),
                           bbox)
nc
# Dimensions of the object are [variable, x, y, date]
# Let's plot the first time steps for the wind speed
plot(nc["sfcWindAdjust",,,1:66])
}
}
