% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drias_list.R
\name{drias_list_remote}
\alias{drias_list_remote}
\alias{drias_list}
\title{List of netCDF file from DRIAS2020/Explore2 database}
\usage{
drias_list_remote(
  url = file.path(cfg$climato$drias$url, cfg$climato$drias$folders),
  nc_pattern = ".*\\\\.nc",
  method = "(ADAMONT|CDFt)",
  ignore_pattern = sprintf("(\%s)", paste(cfg$climato$drias$ignore_pattern, collapse =
    "|")),
  cfg = loadConfig()
)

drias_list(cfg = loadConfig(), with_variables = FALSE, safran = TRUE)
}
\arguments{
\item{url}{\link{character} URLs where to search NetCDF files}

\item{nc_pattern}{Pattern}

\item{method}{\link{character} the bias correction method ("ADAMONT", "CDFt", or "(ADAMONT|CDFt)" for both)}

\item{ignore_pattern}{\link{character} regex pattern for files or path to ignore}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for details}

\item{with_variables}{\link{logical}, if \code{TRUE}, return the same output as \link{drias_list_remote}}

\item{safran}{\link{logical}, if \code{TRUE}, add a row for SAFRAN database}
}
\value{
A \link{data.frame} with the following columns:
\itemize{
\item for \code{drias_list_remote}: GCM, RCM, scenario, time_step, parameter, file, url
\item for \code{drias_list}: if called with argument \code{with_variables=FALSE}, one row by scenario,
and columns GCM, RCM, bias\_correction, scenario, gcm\_rcm, and file. If called
with argument \code{with_variables=FALSE}, same output as \code{drias_list_remote}
except that rows containing "_OLD" are filtered and with extra columns:
ETP ("FAO" or "Hg0175"), and bias\_correction ("ADAMONT" or "CDFt")
}
}
\description{
\code{drias_list} returns the list of the netCDF files processed for the Aude basin.
\code{drias_list_remote} returns the list of netCDF files available on the CNRM server
for the whole metropolitan France territory.
}
\details{
This function is used by the script \code{data-raw/drias_GCM_RCM_scenario_variables.R}
for creating the file \code{drias_GCM_RCM_scenario_variables.tsv} stored on Nextcloud.
}
\examples{
# The function fetches the file list from the CNRM server...
\dontrun{
df <- drias_list_remote()
str(df)
}

}
