% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,stars,sf-method}
\alias{extract,stars,sf-method}
\title{Extract values from a raster \code{stars} object to a vector \code{sf} object}
\usage{
\S4method{extract}{stars,sf}(x, y)
}
\arguments{
\item{x}{SpatRaster or SpatVector of polygons}

\item{y}{SpatVector (points, lines, or polygons). Alternatively, for points, a 2-column matrix or data.frame (x, y) or (lon, lat). Or a vector with cell numbers}
}
\value{
A vector \code{stars} object with extracted values from raster with an
attribute "sf_data" which is a copy of the data contained in \code{y}.
}
\description{
Compute mean area-weighted values of intersections between raster and vector.
}
