% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reaches.R
\name{extract_reaches}
\alias{extract_reaches}
\title{Extract River Reaches by Sub-basin from a DEM}
\usage{
extract_reaches(
  sb,
  d8fd = attr(sb, "files")["d8"],
  streams = attr(sb, "files")["streams"],
  max_tries = 1e+05,
  eps = 1,
  max_step_back = 10,
  max_azimut = 4,
  max_buffer = 3,
  dTolerance = 2 * terra::res(d8fd),
  future_plan = future::multisession,
  future_workers = get_future_workers(),
  filename = getCachePath(sb, "reaches", "gpkg"),
  overwrite = FALSE
)
}
\arguments{
\item{sb}{An \code{\link[sf:sf]{sf}} object representing the perimeter of sub-basins.
Typically the output of \code{\link[=extract_sub_basins]{extract_sub_basins()}}.}

\item{d8fd}{A \code{\link[terra:SpatRaster-class]{SpatRaster}} or \link{character} path to the file containing D8
flow directions.
Can also be obtained from \code{\link[=extract_sub_basins]{extract_sub_basins()}}.}

\item{streams}{An \code{\link[sf:sf]{sf}} object or \link{character} path to the file containing
the stream network
(e.g., output of \code{\link[=get_ign_streams]{get_ign_streams()}}).}

\item{max_tries}{\link{integer}. Maximum cells tested for retrieving a reach.}

\item{eps}{\link{numeric}. Minimum acceptable distance between the downstream reach
and the downstream outlet in cells.}

\item{max_step_back}{\link{integer} Maximum contiguous back steps in stream cell research}

\item{max_azimut}{\link{integer} Maximum change of orientation in cell research
(See details)}

\item{max_buffer}{\link{integer} Maximum number of cells matching around stream cells}

\item{dTolerance}{\link{numeric}. The tolerance for simplifying the reach geometry (
Used in \link[sf:geos_unary]{sf::st_simplify}).}

\item{future_plan}{\link{function} One of available future plan: \link[future:sequential]{future::sequential},
\link[future:multisession]{future::multisession}, \link[future:multicore]{future::multicore}, \link[future:cluster]{future::cluster}}

\item{future_workers}{\link{integer} Number of workers to use for parallel processing.}

\item{filename}{\link{character} Optional. If provided, the streams geometry will
be saved to this file and can be used as cache if \code{overwrite = FALSE}.}

\item{overwrite}{\code{logical}; if TRUE, file is overwrite.}
}
\value{
An \code{\link[sf:sf]{sf}} object containing one river reach per sub-basin,
from the sub-basin outlet to the next downstream outlet.
}
\description{
This function extracts individual river reaches by tracing flow paths from
each sub-basin outlet to the next downstream outlet using a D8 flow direction
raster and a stream network.
}
\details{
Type \code{logger::log_level(logger::DEBUG)} or \code{logger::log_level(logger::TRACE)}
before running the function for getting details on stream cells search.
In this mode, maps zooming on issue zones are recorded in the working
directory (names "\verb{debug_extract_reaches_bvi*_step*_row*_col*.png}").

\code{max_azimut} correspond to the maximum change of orientation when searching
the next stream cell (0, for no change; 2 for 45째; 4 for 90째, 6 for 135째 and
7 for 90째).
}
