% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sub_basins.R
\name{extract_sub_basins}
\alias{extract_sub_basins}
\alias{watershed_get_from_outlet}
\alias{find_near_target}
\title{Extract sub-basins from DEM}
\usage{
extract_sub_basins(
  outlets,
  dem = NULL,
  streams,
  crs = NULL,
  threshold = 1000,
  snap_dist = 100,
  area_tolerance = 0.05,
  max_radius = 50,
  filename = TRUE,
  overwrite = FALSE,
  verbose_mode = NULL
)

watershed_get_from_outlet(
  i,
  outlets,
  file_d8,
  r_flow_accum,
  crs,
  area_tolerance = 0.05,
  max_radius = 50,
  verbose_mode = NULL,
  overwrite = FALSE
)

find_near_target(
  r_flow_accum,
  row0,
  col0,
  target,
  area_tolerance,
  max_radius = 5
)
}
\arguments{
\item{outlets}{\link{data.frame} with columns "id", "x", and "y" and optionnaly "area"
containing respectively the outlet ID, X and Y coordinates (in the CRS of the DEM)
and the a priori area of the sub-basin (in square meters) (See Details)}

\item{dem}{\link{character} containing the filename of the DEM raster, or
a \code{\link[terra:SpatRaster-class]{SpatRaster}} object}

\item{streams}{A \code{\link[sf:sf]{sf}} object containing the streams.}

\item{crs}{The CRS used for outlets coordinates. If \code{NULL}, the CRS of \code{dem}
is used}

\item{threshold}{Minimum threshold of contributing area to extract streams
(unit: square meters)}

\item{snap_dist}{Maximum distance between outlet and a stream vector for
relocating outlet in the stream (in meters)}

\item{area_tolerance}{Maximum relative error between the a priori area
and the computed area of the sub-basin}

\item{max_radius}{Maximum search radius (in number of cells) to find a closer point}

\item{filename}{\link{character} Optional. If provided, the sub-basins geometry will
be saved to this file and can be used as cache if \code{overwrite = FALSE}.
If \code{TRUE} (default), a filename will be automatically generated with
\link{getCachePath} based on the input parameters.}

\item{overwrite}{\code{logical}; if TRUE, file is overwrite.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{i}{\link{integer} row number to process in \code{outlets}}

\item{file_d8}{Path of raster D8}

\item{r_flow_accum}{Raster of flow accumulation}

\item{row0, col0}{respectively row and column number of the current coordinates
of the outlet in the raster}

\item{target}{Target area to reach (in square meters)}
}
\value{
\code{extract_sub_basins} returns a \code{\link[sf:sf]{sf}} object with the
polygons representing the sub-basins and various attributes detailling the
sub-basin up-down order, total basin and sub-basin areas, coordinates of
outlets after move.
}
\description{
Extract sub-basins from DEM
}
\details{
\code{extract_sub_basins} delineate the watershed from the DEM for each outlet,
then it computes the sub-basins contour by remove overlapping basin areas and
compute which sub-basin is downstream of which other sub-basin.

\code{watershed_get_from_outlet} delineate one basin given outlet coordinates.
One can provide an a priori area in \code{outlets$area} which is used to control
the area of the resulting polygon. If the resulting area is outside the tolerance
(\code{area_tolerance}), the algorithm search for a closer point (at a maximum distance
of \code{max_radius} cells of the DEM raster) on the stream network and delineate
again the basin.
This is repeated until the area is within the tolerance or \code{max_tries} is reached.
}
\section{Functions}{
\itemize{
\item \code{extract_sub_basins()}: extracts sub-basins from a DEM using whitebox tools.

\item \code{watershed_get_from_outlet()}: Extract one watershed for one outlet

\item \code{find_near_target()}: Find a cell near (row0, col0) in the flow
accumulation raster cells with a value close to a target area.

}}
