% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_safran.R
\name{get_climate_safran}
\alias{get_climate_safran}
\title{Extract and Aggregate SIM2 Daily Climate Data from RADIS}
\usage{
get_climate_safran(
  x,
  fields = list(Precip = c("PRENEI_Q", "PRELIQ_Q"), PotEvap = "ETP_Q", TempMean = "T_Q"),
  date_start,
  date_end,
  overlay_mode = "aggregate",
  filename = getCachePath(list(x, fields, date_start, date_end, overlay_mode),
    "climate_safran", ifelse(overlay_mode \%in\% c("intersect_geometry", "aggregate"),
    "nc", "RDS")),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{sf} object or geometry supported by \code{RADIS::get_sim2_daily()},
typically representing a spatial feature (e.g., watershed or point).}

\item{fields}{A named list where each element corresponds to a GR-compatible
variable name (e.g., \code{Precip}, \code{PotEvap}, \code{TempMean}) and contains one or
more SIM2 field names to be retrieved and optionally summed.}

\item{date_start}{A \code{Date} object specifying the start of the extraction period.}

\item{date_end}{A \code{Date} object specifying the end of the extraction period.}

\item{overlay_mode}{A character string passed to \code{RADIS::get_sim2_daily()},
typically \code{"aggregate"} for spatial mean over polygons or \code{"raw"} for full raster.}

\item{filename}{If not \code{NULL}, path of cached data (See details).}

\item{overwrite}{\code{logical}; if TRUE, file is overwrite.}

\item{...}{Additional arguments passed to \code{RADIS::get_sim2_daily()}.}
}
\value{
A named list where each element is a 3D array (or similar structure) containing
the extracted and aggregated climate variables, named using the keys of \code{fields}
(e.g., \code{"Precip"}, \code{"PotEvap"}, etc.).
}
\description{
This function retrieves selected daily climate variables from the SIM2 dataset
via the \code{RADIS::get_sim2_daily()} function, renames the fields for compatibility
with GR models (e.g., \code{airGR}), and optionally aggregates multiple related
source fields (e.g., solid and liquid precipitation).
}
\details{
Default value for \code{filename} depends on the parameter \code{overlay_mode} which
cache the data as a netCDF file for \code{overlay_mode = "intersect_geometry"} and
in RDS format for \code{overlay_mode = "intersect_geometry"}.
}
\examples{
\dontrun{
library(sf)
# Example with a watershed polygon
basin <- sf::st_read("my_watershed.gpkg")
data <- get_climate_safran(
  x = basin,
  date_start = as.Date("2000-01-01"),
  date_end = as.Date("2005-12-31")
)
}

}
\seealso{
\code{\link[RADIS:get_sim2_daily]{RADIS::get_sim2_daily()}}
}
