% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_workers.R
\name{get_future_workers}
\alias{get_future_workers}
\title{Compute number of workers given available memory and CPUs}
\usage{
get_future_workers(mem_share = 0.8, cpu_share = 0.8)
}
\arguments{
\item{mem_share}{\link{numeric} Share of RAM to use given total RAM and currently
used RAM by this R session.}

\item{cpu_share}{\link{numeric} Share of CPUs to use given total available CPUs.}
}
\value{
\link{integer} Number of workers to use.
}
\description{
This function computes the number of workers to use for parallel processing
taking into account both constraints that are available RAM and CPUs.
}
\examples{
get_future_workers()
}
