% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hubeau_flows.R
\name{get_hubeau_flows}
\alias{get_hubeau_flows}
\alias{get_hubeau_flows.default}
\alias{get_hubeau_flows.SubBasins}
\alias{get_hubeau_flows.character}
\title{Retrieve river flow observation from the Hydrometrie Hub'Eau API}
\usage{
get_hubeau_flows(x, ...)

\method{get_hubeau_flows}{default}(x, ...)

\method{get_hubeau_flows}{SubBasins}(x, ...)

\method{get_hubeau_flows}{character}(
  x,
  date_start,
  date_end,
  grandeur_hydro_elab = "QmnJ",
  max_days = 10000,
  path_cache = getCachePath(list(date_start, date_end, grandeur_hydro_elab),
    "hubeau_flows"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{Either a \link{character} vector of station codes or a \code{\link[=extract_sub_basins]{SubBasins}} object.}

\item{...}{Parameters passed to methods or to \link[hubeau:get_hydrometrie]{hubeau::get_hydrometrie_obs_elab}}

\item{date_start}{Date of starting period}

\item{date_end}{Date of ending period}

\item{grandeur_hydro_elab}{Processed hydrometric variable type
("HIXM", "HIXnJ", "QINM", "QINnJ", "QixM", "QIXnJ", "QmM", or "QmnJ")}

\item{max_days}{Maximum number of days in one API call}

\item{path_cache}{If not \code{NULL}, folder path of cached data}

\item{overwrite}{\code{logical}; if TRUE, file is overwrite.}
}
\value{
A \link{data.frame} with a first column \code{DatesR} containing the dates and
one column by station containing observed flows in m3/s.
The returned object contains an attribute "codes_site_station" listing the
correspondance between \code{code_entite} and station codes.
}
\description{
Flow observations are retrieved with \link[hubeau:get_hydrometrie]{hubeau::get_hydrometrie_obs_elab} and
formatted for \strong{airGRiwrm} usage.
Station identifiers are used as \code{code_entite} filter parameter on the
Hydrometrie Hub"Eau API and can be either a site (8 characters-length) or a
measuring station (10 characters-length) code.
}
\details{
Hub'Eau APIs are currently limited to 20000 rows so this function cut the time
period in \code{max_days} chunks for the API calls.
}
