% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ign_dem.R
\name{get_ign_dem}
\alias{get_ign_dem}
\alias{get_ign_dem.sf}
\alias{get_ign_dem.bbox}
\alias{get_ign_dem.sfc}
\title{Extract DEM from IGN API}
\usage{
get_ign_dem(x, ...)

\method{get_ign_dem}{sf}(x, ...)

\method{get_ign_dem}{bbox}(x, ...)

\method{get_ign_dem}{sfc}(
  x,
  layer = "ELEVATION.ELEVATIONGRIDCOVERAGE",
  res = 25,
  crs = sf::st_crs(x),
  neg_outlier_threshold = 0,
  ...
)
}
\arguments{
\item{x}{A bounding box, a simple feature collection or a simple feature geometry.}

\item{...}{Further parameters passed to methods and to \link[happign:get_wms_raster]{happign::get_wms_raster}}

\item{layer}{\code{character}; layer name obtained from
\code{get_layers_metadata("wms-r")} or the
\href{https://geoservices.ign.fr/services-web-experts}{IGN website}.}

\item{res}{\code{numeric}; resolution specified in the units of
the coordinate system (e.g., meters for EPSG:2154, degrees for EPSG:4326).
See details for more information.}

\item{crs}{\code{numeric}, \code{character}, or object of class \code{sf} or \code{sfc};
defaults to EPSG:2154. See \code{\link[sf:st_crs]{sf::st_crs()}} for more details.}

\item{neg_outlier_threshold}{Threshold under which data is rejected as outlier
and replaced by the mean of neighbour cells}
}
\value{
A \code{\link[terra:SpatRaster-class]{SpatRaster}} object containing the DEM data.
}
\description{
This function download DEM from IGN API and perform correction on outliers.
}
\details{
Default value of \code{layer} correspond to the BD Alti with 25 m of resolution.
Available DEM on IGN API can be explored with the instruction:
\code{happign::get_layers_metadata("wms-r", "altimetrie")}.
}
\examples{
# Basic usage with default DEM
bbox <- c(xmin = 325522.8, ymin = 6250949.2, xmax = 347137.1, ymax = 6270285.1)
dem <- get_ign_dem(sf::st_bbox(bbox, crs = 2154))
dem
terra::plot(dem)
}
