% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ign_streams.R
\name{get_ign_streams}
\alias{get_ign_streams}
\alias{get_ign_streams.sf}
\alias{get_ign_streams.bbox}
\alias{get_ign_streams.sfc}
\title{Get streams from IGN}
\usage{
get_ign_streams(x, ...)

\method{get_ign_streams}{sf}(x, ...)

\method{get_ign_streams}{bbox}(x, ...)

\method{get_ign_streams}{sfc}(
  x,
  layer = "BDCARTO_V5:troncon_hydrographique",
  crs = sf::st_crs(x),
  filters = NULL,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{A bounding box, a simple feature collection or a simple feature geometry.}

\item{...}{Further parameters passed to methods and to \link[happign:get_wfs]{happign::get_wfs}}

\item{layer}{\code{character}; name of the layer from \code{get_layers_metadata("wfs")}
or directly from
\href{https://geoservices.ign.fr/services-web-experts}{IGN website}}

\item{crs}{The CRS to use for the output streams (the CRS of \code{x} is used by default).}

\item{filters}{A list of filters to apply to the streams data (See details)}

\item{filename}{\link{character} Optional. If provided, the streams geometry will
be saved to this file and can be used as cache if \code{overwrite = FALSE}.}

\item{overwrite}{\code{logical}; if TRUE, file is overwrite.}
}
\value{
A \code{\link[sf:sf]{sf}} object containing the streams data.
}
\description{
This function retrieves hydrographic streams from the IGN WFS service and
filter them based on specified criteria.
}
\details{
List of filters is structured as follow: name correspond to the column names
in the streams data, and values can be:
\itemize{
\item A character vector to filter by specific values
\item A function that takes a vector and returns a logical vector (e.g., \code{not.is.na})
\code{not.is.na} is equivalent of \code{!is.na()}.
}
}
\examples{
bbox <- c(xmin = 325522.8, ymin = 6250949.2, xmax = 347137.1, ymax = 6270285.1)
# Retrieve only river streams that have a name and of specific nature
streams <- get_ign_streams(
  sf::st_bbox(bbox, crs = 2154),
  filters = list(
    nature = c("Ecoulement naturel", "Ecoulement canalis\u00E9"),
    cpx_toponyme_de_cours_d_eau = not.is.na
  )
)
streams
mapview::mapview(streams)
}
