% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_safran_var_units.R
\name{get_safran_var_units}
\alias{get_safran_var_units}
\title{Get units of the variables available in SAFRAN database}
\usage{
get_safran_var_units()
}
\value{
named \link{character} \link{vector} containing units for each climatic variable
}
\description{
Get units of the variables available in SAFRAN database
}
\examples{
get_safran_var_units()
}
