% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCachePath.R
\name{getCachePath}
\alias{getCachePath}
\alias{getCacheDir}
\alias{clearCache}
\title{Get cache path}
\usage{
getCacheDir(
  pkg = utils::packageName(),
  basepath = Sys.getenv("CACHE_DIR_AIRGRCCIA", dirname(tempdir())),
  pattern = "\%s_cache"
)

getCachePath(
  x,
  prefix = as.character(substitute(x)),
  fileext = NULL,
  cache_dir = getCacheDir()
)

clearCache(cache_dir = getCacheDir())
}
\arguments{
\item{pkg}{Package name (default current package)}

\item{basepath}{Folder in which to add cache directory (default temp)}

\item{pattern}{Pattern applied on package name with \link{sprintf}}

\item{x}{Object to save (or property of object that can be used as a signature)}

\item{prefix}{Optional prefix to add to the file name (default variable name)}

\item{fileext}{Optional file extension to use for the file (e.g., "rds", "csv", etc.)}

\item{cache_dir}{Path to the cache directory (default from \code{\link[=getCacheDir]{getCacheDir()}})}
}
\value{
The cache path.
}
\description{
Build a path in a cache directory and a file name build from MD5 hash.
}
\section{Functions}{
\itemize{
\item \code{getCacheDir()}: Returns cache path root and create the folder if it
does not exist.

\item \code{getCachePath()}: Returns cache path for a specific file.

\item \code{clearCache()}: Clear the cache directory

}}
\examples{
getCacheDir()
path_cars <- getCachePath(cars, fileext = "rds")
path_cars
\dontrun{
saveRDS(cars, path_cars)
}
}
