% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_get_stations.R
\name{hydro_get_station}
\alias{hydro_get_station}
\alias{hydro_station_as_sf}
\title{Get info on hydrometric station}
\usage{
hydro_get_station(x)

hydro_station_as_sf(x)
}
\arguments{
\item{x}{code site}
}
\value{
A \link{data.frame} or a SpatVector. See \link[hubeau:get_hydrometrie]{hubeau::get_hydrometrie_sites}.
}
\description{
Get info on hydrometric station
}
\examples{
hydro_get_station("Y2300020")
hydro_station_as_sf("Y2300020")
}
