% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ign_dem.R
\name{is_projected_crs}
\alias{is_projected_crs}
\title{Check if the CRS is projected}
\usage{
is_projected_crs(x)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf}}, \code{\link[sf:sfc]{sf}} geometry, or \code{\link[sf:st_bbox]{bbox}}.}
}
\value{
\code{TRUE} if the CRS is projected, \code{FALSE} otherwise.
}
\description{
Check if the CRS is projected
}
\examples{
# Define the bounding box coordinates in WGS84
bounding_box <- c(xmin = 3.211086, ymin = 43.24153, xmax = 3.904615, ymax = 44.28818)

# Create the bounding box using st_bbox
bbox <- sf::st_bbox(bounding_box, crs = 4326)

# Is it in projected CRS?
is_projected_crs(bbox)

# Convert to projected CRS
bbox_projected <- sf::st_transform(bbox, 2154)
is_projected_crs(bbox_projected)
}
