% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ign_streams.R
\name{not.is.na}
\alias{not.is.na}
\title{Shortcut for \code{!is.na}}
\usage{
not.is.na(x)
}
\arguments{
\item{x}{an R object to be tested}
}
\value{
\link{logical} See \link{is.na}.
}
\description{
Shortcut for \code{!is.na}
}
