% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_uncertainty.R
\name{predict_uncertainty_calibration}
\alias{predict_uncertainty_calibration}
\alias{predict_uncertainty}
\alias{plot_uncertainty_time_series}
\alias{plot_uncertainty_calibration}
\title{Predictive Uncertainty functions}
\usage{
predict_uncertainty_calibration(
  Qobs,
  Qsim,
  bind_size = 0.1,
  bind_step = 0.01,
  FUN_TRANS = log10,
  confidence_interval = 0.9,
  probs = c((1 - confidence_interval)/2, 1 - (1 - confidence_interval)/2)
)

predict_uncertainty(Qsim, uncertainty_table)

plot_uncertainty_time_series(
  Qsim_uncertain,
  DatesR,
  Qobs = NULL,
  bounds = names(Qsim_uncertain)[c(2, ncol(Qsim_uncertain))],
  rubbon_fill = "darkblue",
  rubbon_alpha = 0.2,
  line_colors = c(Simulated = "orangered", Observed = "black"),
  labels = list(title = "Predicted Discharge with Uncertainty Intervals", x = "Date", y =
    "Discharge (mm)", color = "Flow type")
)

plot_uncertainty_calibration(
  Qsim_uncertain,
  Qobs,
  bounds = names(Qsim_uncertain)[c(2, ncol(Qsim_uncertain))],
  rubbon_fill = "darkblue",
  rubbon_alpha = 0.2,
  outliers_quantile = 0.999,
  labels = list(title =
    "Relative Error vs. Simulated Discharge with Uncertainty Intervals", x =
    "Simulated Discharge (mm)", y = "Relative Error Qobs/Qsim (-)")
)
}
\arguments{
\item{Qobs}{Observed discharge values (numeric vector).}

\item{Qsim}{Simulated discharge values (numeric vector).}

\item{bind_size}{Size of the moving bin as a fraction of the data range (default is 0.1).}

\item{bind_step}{Step size for moving the bin as a fraction of the data range (default is 0.01).}

\item{FUN_TRANS}{Transformation function to apply to discharge values (default is log10).}

\item{probs}{Probabilities for quantile calculation (default is c(0.05, 0.5, 0.95))
for a 90\% prediction interval.}

\item{DatesR}{Dates corresponding to the discharge values.}

\item{bounds}{Names of the columns in \code{QSim_uncertain} representing the uncertainty bounds (default is the first and last error quantile).}

\item{labels}{List of labels for the plot (title, x, y, ...) used in
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{outliers_quantile}{Quantile threshold to filter out extreme relative
errors (default is 0.999).}

\item{QSim_uncertain}{Data frame returned by \code{predict_uncertainty}, containing simulated discharge and uncertainty quantiles.}
}
\value{
A data frame with columns:
\itemize{
\item \code{Qmin}: Minimum discharge value of the bin.
\item \code{Qmax}: Maximum discharge value of the bin.
\item Quantiles of the relative error (Qsim / Qobs) for the specified probabilities.
}

Each row corresponds to a bin defined by \code{bind_size} and \code{bind_step}.
}
\description{
Predictive Uncertainty functions
}
\details{
This method is based on the approach described in Bourgin (2014) pages 201-203
with a variant consisting in moving a bin along the distribution of
simulated discharge instead of fixed bins.

References:

Bourgin, François. 2014. « Comment quantifier l’incertitude prédictive en
modélisation hydrologique ? : Travail exploratoire sur un grand échantillon
de bassins versants ». Phdthesis, AgroParisTech.
https://pastel.archives-ouvertes.fr/tel-01130084.
}
\section{Functions}{
\itemize{
\item \code{predict_uncertainty_calibration()}: Predictive Uncertainty Calibration

Model of predictive uncertainty based on quantiles of the empirical
distribution of relative errors for different bins of simulated discharge
values using a moving bin approach.

\item \code{predict_uncertainty()}: Predictive Uncertainty Prediction
Using the uncertainty table generated by \code{predict_uncertainty_calibration},
this function predicts the relative uncertainty for new simulated discharge values.

\item \code{plot_uncertainty_time_series()}: Plot Predicted Uncertainty Time Series.
This function visualizes the predicted discharge values along with their uncertainty intervals.

\item \code{plot_uncertainty_calibration()}: Plot Predicted Uncertainty Calibration.
This function visualizes the relative error between observed and simulated
discharge values,
along with the predicted uncertainty intervals.

}}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)

data(L0123001, package = "airGR")
InputsModel <- CreateInputsModel(
  RunModel_GR4J,
  DatesR = BasinObs$DatesR,
  Precip = BasinObs$P,
  PotEvap = BasinObs$E
)
Ind_Run <- seq(
  which(format(BasinObs$DatesR, format = "\%Y-\%m-\%d") == "1985-01-01"),
  which(format(BasinObs$DatesR, format = "\%Y-\%m-\%d") == "2012-12-31")
)
RunOptions <- CreateRunOptions(
  RunModel_GR4J,
  InputsModel = InputsModel,
  IndPeriod_Run = Ind_Run,
  IndPeriod_WarmUp = (Ind_Run[1] - 365):(Ind_Run[1] - 1)
)
Param <- c(257.237556, 1.012237, 88.234673, 2.207958)
OutputsModel <- RunModel_GR4J(
  InputsModel = InputsModel,
  RunOptions = RunOptions,
  Param = Param
)
Qsim <- OutputsModel$Qsim
Qobs <- BasinObs$Qmm[Ind_Run]

uncertainty_table <- predict_uncertainty_calibration(Qobs, Qsim)


Qsim_uncertain <- predict_uncertainty(Qsim, uncertainty_table)

# Plot the result of uncertainty estimation
plot_uncertainty_calibration(Qsim_uncertain, Qobs)

# Plot the time series with uncertainty intervals
ts_selection <- 1:365
plot_uncertainty_time_series(
  Qsim_uncertain[ts_selection, ],
  BasinObs$DatesR[Ind_Run[ts_selection]]
)

# Plot the time series with uncertainty intervals compared with observations
plot_uncertainty_time_series(
  Qsim_uncertain[ts_selection, ],
  BasinObs$DatesR[Ind_Run[ts_selection]],
  Qobs[ts_selection]
)
}
