% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drias_crop_aggregate.R
\name{read_drias}
\alias{read_drias}
\title{Read a NetCDF file from DRIAS2020 database}
\usage{
read_drias(
  url = NULL,
  filename = basename(url),
  path = tempdir(),
  crs = "EPSG:27572"
)
}
\arguments{
\item{url}{\link{character} URL of the file for downloading}

\item{filename}{\link{character} Name of the file to download}

\item{path}{\link{character} Local path of the downloaded file}

\item{crs}{\link{character} for replacing the \code{crs} attributes of the \code{stars} object}
}
\value{
A \code{stars} object (See stars::read_mdim)
}
\description{
This function downloads a netCDF file, open it with \link[stars:mdim]{stars::read_mdim},
correct the CRS by setting it to Lambert II , and returns
the modified stars object.
}
\details{
The undefined CRS that is corrected here is due to a wrong format of the netCDF
or a bug in GDAL, see https://github.com/r-spatial/stars/issues/628#issuecomment-1537946005
for details.
}
\examples{
\dontrun{
s <- read_drias("https://climatedata.umr-cnrm.fr/public/dcsc/projects/DRIAS/DRIAS2020_NEW/CNRM-CERFACS-CNRM-CM5/CNRM-ALADIN63/historical/day/prtotAdjust/prtotAdjust_France_CNRM-CERFACS-CNRM-CM5_historical_r1i1p1_CNRM-ALADIN63_v2_MF-ADAMONT-SAFRAN-1980-2011_day_19510101-20051231.nc")
s
}
}
