% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_to_outlets.R
\name{sb_to_outlets}
\alias{sb_to_outlets}
\title{Title}
\usage{
sb_to_outlets(sb)
}
\arguments{
\item{sb}{Output of \link{extract_sub_basins}.}
}
\value{
A simple feature with points representing sub-basins outlets.
}
\description{
Title
}
